pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[project]
name = "python-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.4",
    "fastapi>=0.128.8",
    "inngest>=0.4.22",
    "langchain-text-splitters>=1.1.1",
    "llama-index-core>=0.14.16",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.30.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.1",
    "ragas>=0.4.3",
    "streamlit>=1.50.0",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
